requirements.txt
pyTelegramBotAPI
asyncio
aioeos
redis
pycryptodome
requests
ueosio
python-dotenv